chore: move some more stuff around
